LPD-53339 Create Paginated Table component
